pull helper methods out into their own response_helpers module
